requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0

# OpenAI API
openai==1.3.7

# Firebase dependencies
firebase-admin==6.2.0
google-cloud-firestore==2.13.1

# Audio processing
pydub==0.25.1

# Utilities
requests==2.31.0
werkzeug==2.3.7
gunicorn==21.2.0

# Date/time handling
python-dateutil==2.8.2

# JSON handling
jsonschema==4.19.2